requirement.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.8.1
async-generator==1.10
async-timeout==4.0.2
branca==0.7.2
cachetools==5.3.0
certifi==2024.2.2
cfgv==3.3.1
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
coverage==7.2.3
distlib==0.3.6
Django==4.2.11
dnspython==2.3.0
docker==6.0.1
docstring-to-markdown==0.12
duckduckgo-search==2.8.6
exceptiongroup==1.1.1
exif==1.6.0
flake8==6.0.0
flatbuffers==23.3.3
folium==0.16.0
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gTTS==2.3.1
h11==0.14.0
httplib2==0.22.0
identify==2.5.22
idna==3.7
Jinja2==3.1.3
loguru==0.7.0
MarkupSafe==2.1.5
mccabe==0.7.0
mediapipe==0.9.1.0
multidict==6.0.4
nodeenv==1.7.0
numpy==1.24.4
openai==0.27.2
opencv-contrib-python==4.7.0.72
orjson==3.8.10
outcome==1.2.0
packaging==23.0
pillow==10.3.0
pinecone-client==2.2.1
playsound==1.2.2
plum-py==0.8.7
pre-commit==3.2.2
protobuf==3.20.3
pycodestyle==2.10.0
pydocstyle==6.2.3
pyflakes==3.0.1
python-dotenv==1.0.0
python-lsp-server==1.7.1
pytoolconfig==1.2.5
readability-lxml==0.8.1
redis==4.5.4
requests==2.28.2
rope==1.7.0
rsa==4.9
selenium==4.8.3
sourcery==1.2.0
sqlparse==0.5.0
tiktoken==0.3.3
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
virtualenv==20.21.0
webdriver-manager==3.8.6
whatthepatch==1.0.4
win32-setctime==1.1.0
wsproto==1.2.0
xyzservices==2024.4.0
yarl==1.8.2
